WIP: Fusion Parser Cached
Idea: don't reparse all fusion files over and over. Cache the partial ast somehow and invalidate those ast partials when a file changed, then set those ast partials together, on a parse.

## Usage:

use it like the normal parser:
(Caching is already configured.)
`(new CachedParser())->parse($mergedFusionCode, $siteRootFusionPathAndFilename);`

------

## Solved challenges:
#### how to identify a fusion ast from cache and how to identify changes of a file and ast?
By the required $contextPathAndFilename - except not really since this is unresolved (could include different './../') and is most likely a resource wrapper. The problem is, that we need to normalize the paths, so we can compare them with the monitored files that changed. So we must resolve the file via realpath - but realpath doesn't work on stream wrapper like 'resource://' so we used the same technik as in the RessourceStreamWrapper to get the absolute path (with the flow root etc.) and call realpath on this.

#### How to separate the fusion ast from its includes?
We override the method of the parser which would resolve the file pattern and parse those fusion files and add them to the ast. Instead, we implement an observer/watcher which will remember the includes, so we can resolve them separately and also cache the raw include patterns with the ast together. This is especially important as when there is a glob include, and a new file.fusion is added, we want to include this file too. So raw file patterns are cached to allow globing.

#### How to unset a path `root` in a file in which it was not declared?
Simply parsing one file isolated won't do - since unsets `>` will only work this way per-file basis and not like in the real parser - on all previous parsed fusion. To make this work correctly, we remember each unset path of a file and later unset those paths separately.

#### set a path to null.
doesnt work anymore with 67a8af87add0b8d3dad9e6d20a0c82e387fce217

------

## Demo

This way we would get for this simple fusion, following cache entry:

**$contextPathAndFilename:**
`resource://Vendor.Site/Private/Fusion/Root.fusion`

**$sourceCode:**
```
include: **/*.fusion

root >
root = "foo"
```

```php
[
    /* 1) include */
    ['**/*.fusion'],
    /* 2) fusion */
    [
        'root' => [
            '__stopInheritanceChain' => true,
            '__value' => 'foo',
            '__eelExpression' => NULL,
            '__objectType' => NULL,
        ],
    ]
    /* 3) unset paths */
    [
        ['root']
    ],
]
```

the cache entry id is a hash from the resolved absolute realpath (with removed flow root) of the $contextPathAndFilename `resource://Vendor.Site/Private/Fusion/Root.fusion` will become:
`md5("Packages/Sites/Neos.Demo/Resources/Private/Fusion/Root.fusion")`

------

## Unresolved challenges:
#### `include:` everywhere
`include:` currently works at every position in a fusion file (unlike in php the include is not a copy and paste but the ast is merged at first level.) The cached parser treats all includes like if they were at the top - that could result in a different ast - but as discussed includes everywhere don't make sense.

#### copy path feature disabled (prototype inheritance not affected!)
a simple `foo < somepath` doesn't work with this cached parser and will throw an exception. For this to work it must be evaluated separately to work over multiple files. I would be fine with not supporting this and then using the normal parser. We do use still on copy: https://github.com/neos/neos-development-collection/blob/939da092b86935aece4e0231464b3f0d41f4591f/Neos.NodeTypes.Navigation/Resources/Private/Fusion/Root.fusion#L15 - Sure one could also allow this but i wanted to keep the code small at first.