TASK: `NodeIdentity` should contain `WorkspaceName` instead of the `ContentStreamId`
For most, if not all use cases, like caching a reference to a node in fusions @cache.context or serializing a node into the frontend and passing it around as query parameter, the workspace name should be used.

That guarantees that a node is still findable, when a workspace will be rebased (which _can_ happen anytime, and thus the csid would be changed.